scripts: hid_configurator: Move module configuration to separate files
Change moves module configuration related code to separate files.

Jira:DESK-593
