+++ scripts/hid_configurator/configurator_cli.py
@@ -6,9 +6,9 @@
 import argparse
 import logging
 
-from configurator_core import DEVICE
-from configurator_core import get_device_pid, open_device
-from configurator_core import change_config, fetch_config
+from configurator_core import open_device
+from devices import DEVICE, get_device_pid
+from modules.config import change_config, fetch_config
 from modules.dfu import fwinfo, fwreboot, dfu_transfer, get_dfu_image_version
 from modules.led_stream import send_continuous_led_stream
 from modules.music_led_stream import send_music_led_stream
